staking page add multiple assets
